Refactor ThermodynamicSeaIceModel into a new module EulerianThermodynamicSeaIceModels
